Refactor design - 1. Flatten + Abstract 2. LLMError Refactoring 3. Safe Migration - Old LLMError to New LLMError
